Refactored to not need a Pickler
