+++ core/js/src/main/scala/profig/PlatformPickler.scala
@@ -1,4 +1,5 @@
 package profig
 
 trait PlatformPickler {
+  def initProfig(loadModules: Boolean): Unit = {}
 }
