--- core/js/src/main/scala/profig/package.scala
@@ -1,7 +1,0 @@
-import scala.concurrent.Future
-
-package object profig extends SharedJSONConversions {
-  def initProfig(loadModules: Boolean): Unit = {
-    Future.successful(())
-  }
-}
