--- core/jvm/src/main/scala/profig/package.scala
@@ -1,1 +1,0 @@
-package object profig extends ProfigJVMSupport
