+++ core/shared/src/main/scala/profig/JsonUtil.scala
@@ -1,7 +1,5 @@
 package profig
 
-import Pickler._
-
 object JsonUtil {
   def fromJson[T: Reader](json: Json): T = try {
     json.as[T]
