--- core/shared/src/main/scala/profig/SharedJSONConversions.scala
@@ -1,11 +1,0 @@
-package profig
-
-import java.util.Properties
-
-import scala.language.implicitConversions
-
-trait SharedJSONConversions {
-  implicit def properties2JSON(properties: Properties): Json = ProfigUtil.properties2Json(properties)
-  implicit def args2JSON(args: Seq[String]): Json = ProfigUtil.args2Json(args)
-  implicit def json2Value(json: Json): ujson.Value = json.value
-}
